+++ clang-tools-extra/clang-tidy/bsl/LiteralsNoOctalCheck.h
@@ -1,0 +1,42 @@
+//===--- LiteralsNoOctalCheck.h - clang-tidy --------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BSL_LITERALSNOOCTALCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BSL_LITERALSNOOCTALCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace bsl {
+
+/// Warns when non-zero octal literals or escape sequences are used.
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/bsl-literals-no-octal.html
+class LiteralsNoOctalCheck : public ClangTidyCheck {
+public:
+  LiteralsNoOctalCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+  bool isLanguageVersionSupported(const LangOptions &LangOpts) const override {
+    return LangOpts.CPlusPlus;
+  }
+
+private:
+  void checkInteger(const ast_matchers::MatchFinder::MatchResult &Result);
+  void checkString(const ast_matchers::MatchFinder::MatchResult &Result);
+  void checkCharacter(const ast_matchers::MatchFinder::MatchResult &Result);
+};
+
+} // namespace bsl
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BSL_LITERALSNOOCTALCHECK_H